Remove MainViewModel. Change 3 hour PeriodicWorkRequestBuilder to 1 hour
