+++ app/src/main/java/org/covidwatch/android/presentation/MainActivity.kt
@@ -1,52 +1,13 @@
 package org.covidwatch.android.presentation
 
-import android.app.NotificationChannel
-import android.app.NotificationManager
-import android.content.Context
-import android.os.Build
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
-import androidx.core.app.NotificationManagerCompat
-import org.covidwatch.android.CHANNEL_ID
-import org.covidwatch.android.NotificationFactory
 import org.covidwatch.android.R
-import org.covidwatch.android.presentation.util.EventObserver
-import org.koin.android.ext.android.inject
-import org.koin.androidx.viewmodel.ext.android.viewModel
 
 class MainActivity : AppCompatActivity() {
-
-    private val mainViewModel: MainViewModel by viewModel()
-    private val notificationFactory: NotificationFactory by inject()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
-
-        if (savedInstanceState == null) {
-            createNotificationChannel()
-        }
-        mainViewModel.possibleExposureNotificationEvent.observe(this, EventObserver {
-            notifyUserOfPossibleExposure()
-        })
-    }
-
-    private fun createNotificationChannel() {
-        // Create the NotificationChannel, but only on API 26+ because
-        // the NotificationChannel class is new and not in the support library
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-            val name = getString(R.string.channel_name)
-            val importance = NotificationManager.IMPORTANCE_HIGH
-            val channel = NotificationChannel(CHANNEL_ID, name, importance).apply {
-                description = getString(R.string.channel_description)
-            }
-            val notificationManager = getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
-            notificationManager.createNotificationChannel(channel)
-        }
-    }
-
-    private fun notifyUserOfPossibleExposure() {
-        val notification = notificationFactory.possibleExposure()
-        NotificationManagerCompat.from(this).notify(0, notification)
     }
 }
